pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "onenote-mcp-server"
version = "0.1.0"
description = "MCP server (SSE & Streamable HTTP) to connect to OneNote and list notes and note content"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Developer" }]
dependencies = [
    "mcp[cli]>=1.26.0",
    "msgraph-sdk>=1.0.0",
    "azure-identity>=1.15.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "httpx>=0.27.0",
]

[project.scripts]
onenote-mcp = "onenote_mcp.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
